Wire the notifier to the events the node already publishes
The dispatcher starts after the message service and subscribes to the
hub. Subscribing rather than calling into internal/messages is what keeps
this at zero coupling: the message service already publishes an event
when one arrives, so nothing in that package changes and — the part that
mattered for this unit — the DMR shim is not touched at all. Its contract
is that an observer never mutates or blocks a frame, and nothing here
comes near it.

It is also how weather alerts will arrive. The weather service publishes
hub events too, so when it lands the only change here is another case in
the switch.

An inbound message notifies the station it was addressed to, resolved
through the phonebook by DMR ID. A miss enqueues a broadcast rather than
dropping it: somebody should see that a message arrived, and dropping it
would silently lose the notification for exactly the node that has not
filled its phonebook in.

The notification says a message arrived and names who from. It never
carries the text — the hub event does not either, for the same reason:
correspondence does not belong on a surface the operator never agreed to.

Enqueueing never fails a caller. Every path into it must not wait or fail
— an API handler, or a goroutine draining a channel the hub publishes
into — so a queue error is logged and dropped. A missed email is a missed
email; a caller that blocked on one would be a broken dashboard or a
stalled event stream.

The test-send button enqueues rather than sends, which is the whole
point: a test that delivered inline would hang the dashboard for the SMTP
timeout against exactly the misconfigured server it exists to diagnose.
It answers with the queued id and points at the parked list, so an
operator who sees nothing arrive can read the reason instead of guessing.
It refuses outright for an entry with no address, because a test that
reported success and delivered nothing is worse than no test.

The hub event type is referenced by constant rather than by literal. A
rename in internal/messages would otherwise leave the switch matching
nothing, and a notification that stops being sent is not a failure
anything reports.

Also carries the 32-bit test fix from the auth branch, independently:
this branch is cut from main, where an untyped 5000000000 in the
phonebook tests still does not compile for arm. Without it the
cmd/waypointd test binary cannot be built for the bench box at all — and
the stale binary from a previous build passes, which is worse than
failing.

+++ cmd/waypointd/phonebook_test.go
@@ -222,7 +222,10 @@ func TestPhonebookRejections(t *testing.T) {
 		// The DMR ID cases are the reason the wire type decodes into an int64: a
 		// uint32 field would answer these with a JSON decoder's overflow message
 		// instead of a limit the operator can act on.
-		{"DMR ID too wide", map[string]any{"callsign": "W1AW", "dmr_id": 5000000000}, "dmr_id"},
+		// int64, not an untyped constant: on a 32-bit target (the armv6 the image
+		// ships) an untyped 5000000000 defaults to int and does not fit, so this
+		// file would not compile for the architecture it is meant to protect.
+		{"DMR ID too wide", map[string]any{"callsign": "W1AW", "dmr_id": int64(5000000000)}, "dmr_id"},
 		{"DMR ID zero", map[string]any{"callsign": "W1AW", "dmr_id": 0}, "dmr_id"},
 		{"DMR ID negative", map[string]any{"callsign": "W1AW", "dmr_id": -1}, "dmr_id"},
 	} {
